guardrails-proxy/requirements.txt
guardrails-ai==0.5.0
openai==1.36.0
fastapi==0.110.3
uvicorn==0.30.1
httpx==0.27.0
requests==2.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
nltk==3.9.1